refactor(site): extract useRuns hook for dashboard/prompts pages
Both dashboard-page.tsx and prompts-page.tsx duplicated the same useEffect
fetch pattern with cancellation flag. Extracted to shared useRuns() custom
hook in site/src/app/hooks/useRuns.ts.

Refactored both components to use the hook. No behavior change — cancellation
semantics preserved exactly.

Rebuilt embedded site assets (site/dist/) with changes.

Closes #595
